stt,tts,vad: add timeout functionality
introduces two new named arguments to TTS, STT, and VAD:
`connect_timeout: float` and `keepalive_timeout: float`.
if set to 0, they will be ignored. if not, the abstract
`async __anext__()` gets wrapped in a timeout.
`STT.recognize` is not iterable so its timeout is implemented at the
plugin level.

defaults to 0 to preserve backwards compatibility.