Add support for Azure TTS custom voices through an additional argument (endpoint_id) to the TTS class (AGT-445)
Added an endpoint_id property to the TTS() constructor to allow for passing the Azure endpoint_id.  When combined with the voice property, this allows customers to use a custom voice.

Also updated the exception message to get more details when the speak_text_async() is cancelled.  It was challenging to debug the issue without these details.