Added extra_options parameter to OpenAI and Anthropic plugins
This should have been added by default from the beginning, as we want to have access to all the parameters from the LLM APIs to be able to customize them.

One use case would be observability and monitoring. I tried to integrate LiveKit with Langfuse, and I could pass the additional parameters needed to the completions call.

This PR fixes that and let's other developers customize the LLM settings how they want, instead of waiting for each parameter to be added by the LK team (like parallel_tool_calls or tool_choice).