fix: fix incorrect error case
`_process_stream_event` implicitly assumed that error messages from AssemblyAI had a `message_type` key. The [docs](https://www.assemblyai.com/docs/speech-to-text/streaming) show that this is not the case, and error messages have only the `error` key. This changes reflects this fact and refactors a bit for clarity + robustness.