Bug fix: _play_speech get stuck due to orphan speech handle.
When a new _synthesize_answer_task is created, it is possible that the old _synthesize_answer_task is waiting for the llm_stream:
llm_stream = await llm_stream

In this case, the new task will cancel the old one. However, the handle of the old task is not correct cancelled.

This would result in the handle never being initialized by _synthesize_answer_task.

As a result, the _play_speech function will stuck at
await speech_handle.wait_for_initialization(), and blocks all the following speeches.

Note that the old handle is supposed to be cancelled by self._pending_agent_reply.cancel() in the _synthesize_agent_reply function. However, this is not working properly since self._pending_agent_reply is assigned to None in the _validate_reply_if_possible function.