feat: cgroups v1 support, lower idle processes
the way we get CPU count isn't accurate. on cgroupsv1 instances, the quota
is often unset, leading it to still use the host CPU count.

here we'll put a upper cap on concurrent processes spun up as well as
idle processes, to reduce the likelihood that it'll try to use the entire
machine