fix(voice): Defer user_state transition in STT mode until turn commitment
We've been using `turn_detection="stt"`, and as recommended in the docs we've also been passing a `vad` for interruption handling.

This works nicely 99% of the time, but we had a rare bug where it seemed like the messages from our users were off-by-one. That is, the user would speak, but the agent would answer as if it was always one turn behind.

We traced this back to a race condition:
- Our STT hadn't yet delivered `END_OF_SPEECH`
- But we'd already moved to `user_state: "listening"` based on the VAD
- This allowed the agent to speak. Occasionally, it happened to be ready to do so in this small window.
- Which in turn blocked our processing of the STT `END_OF_SPEECH` until the agent had finished speaking
- Only after which were we able to process the customer's speech
- Which we then generated a response to a turn behind

In both `vad` and `realtime_llm` turn detection, the user_state is aligned with the turn commitment. But when using `stt`, they can fall out of sync like this.

Using the VAD's fast decision is useful for interruption handling at `START_OF_SPEECH`, but at `END_OF_SPEECH` I think it's preferable to use the STT such that the user state can't fall out of sync with turn commitment.